requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.2=py37hecd8cb5_1001
backcall=0.2.0=pyhd3eb1b0_0
basic-image-eda=0.0.3=pypi_0
beautifulsoup4=4.9.3=pypi_0
blis=0.7.4=pypi_0
bottleneck=1.3.2=pypi_0
ca-certificates=2021.1.19=hecd8cb5_0
catalogue=1.0.0=pypi_0
certifi=2020.12.5=py37hecd8cb5_0
chardet=4.0.0=pypi_0
cycler=0.10.0=pypi_0
cymem=2.0.5=pypi_0
decorator=4.4.2=pyhd3eb1b0_0
fastai=1.0.61=pypi_0
fastprogress=1.0.0=pypi_0
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
importlib-metadata=3.4.0=pypi_0
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.19.0=py37h01d92e1_1
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.2=py37hecd8cb5_1
joblib=1.0.0=pypi_0
jupyter_client=6.1.7=py_0
jupyter_core=4.7.0=py37hecd8cb5_0
kiwisolver=1.3.1=pypi_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libsodium=1.0.18=h1de35cc_0
matplotlib=3.3.4=pypi_0
murmurhash=1.0.5=pypi_0
ncurses=6.2=h0a44026_1
numexpr=2.7.2=pypi_0
numpy=1.20.0=pypi_0
nvidia-ml-py3=7.352.0=pypi_0
openssl=1.1.1i=h9ed2024_0
packaging=20.9=pypi_0
pandas=1.2.1=pypi_0
parso=0.7.0=py_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.0=pypi_0
pip=20.3.3=py37hecd8cb5_0
plac=1.1.3=pypi_0
preshed=3.0.5=pypi_0
prompt-toolkit=3.0.8=py_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pygments=2.7.4=pyhd3eb1b0_0
pynvx=1.0.0=pypi_0
pyparsing=2.4.7=pypi_0
python=3.7.9=h26836e1_0
python-dateutil=2.8.1=pyhd3eb1b0_0
pytz=2020.5=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=20.0.0=py37h23ab428_1
readline=8.1=h9ed2024_0
requests=2.25.1=pypi_0
scikit-learn=0.24.1=pypi_0
scipy=1.6.0=pypi_0
seaborn=0.11.1=pypi_0
setuptools=52.0.0=py37hecd8cb5_0
six=1.15.0=py37hecd8cb5_0
sklearn=0.0=pypi_0
soupsieve=2.1=pypi_0
spacy=2.3.5=pypi_0
sqlite=3.33.0=hffcf06c_0
srsly=1.0.5=pypi_0
thinc=7.4.5=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hb0a8c7a_0
torch=1.7.1=pypi_0
torchvision=0.8.2=pypi_0
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.26.3=pypi_0
wasabi=0.8.2=pypi_0
wcwidth=0.2.5=py_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.3=hb1e8313_3
zipp=3.4.0=pypi_0
zlib=1.2.11=h1de35cc_3
